not every value return in an nlist structure was being coerced to an (off_t) which was only being smashed back down to a (u_long) as an argument to kvm_read.  If you ask me nl.n_value ought to be a (void *) but since the type of nl.n_value and the second argument to kvm_read are the same I got fed up and changed all the off_t's to be that type (u_long).
SCCS-vsn: usr.bin/netstat/if.c 5.18
SCCS-vsn: usr.bin/netstat/inet.c 5.19
SCCS-vsn: usr.bin/netstat/iso.c 5.13
SCCS-vsn: usr.bin/netstat/main.c 5.30
SCCS-vsn: usr.bin/netstat/mbuf.c 5.12
SCCS-vsn: usr.bin/netstat/mroute.c 7.2
SCCS-vsn: usr.bin/netstat/netstat.h 5.4
SCCS-vsn: usr.bin/netstat/ns.c 5.16
SCCS-vsn: usr.bin/netstat/route.c 5.25
SCCS-vsn: usr.bin/netstat/unix.c 5.14

+++ usr/src/usr.bin/netstat/mroute.c
@@ -8,7 +8,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)mroute.c	7.1 (Berkeley) %G%
+ *	@(#)mroute.c	7.2 (Berkeley) %G%
  */
 
 /*
@@ -34,7 +34,7 @@
 
 void
 mroutepr(mrpaddr, mrtaddr, vifaddr)
-	off_t mrpaddr, mrtaddr, vifaddr;
+	u_long mrpaddr, mrtaddr, vifaddr;
 {
 	u_int mrtproto;
 	struct mrt *mrttable[MRTHASHSIZ];
@@ -53,7 +53,7 @@ mroutepr(mrpaddr, mrtaddr, vifaddr)
 		return;
 	}
 
-	kread((off_t)mrpaddr, (char *)&mrtproto, sizeof(mrtproto));
+	kread(mrpaddr, (char *)&mrtproto, sizeof(mrtproto));
 	switch (mrtproto) {
 
 	case 0:
@@ -80,7 +80,7 @@ mroutepr(mrpaddr, mrtaddr, vifaddr)
 	saved_nflag = nflag;
 	nflag = 1;
 
-	kread((off_t)vifaddr, (char *)&viftable, sizeof(viftable));
+	kread(vifaddr, (char *)&viftable, sizeof(viftable));
 	banner_printed = 0;
 	for (vifi = 0, v = viftable; vifi < MAXVIFS; ++vifi, ++v) {
 		if (v->v_lcl_addr.s_addr == 0)
@@ -104,7 +104,7 @@ mroutepr(mrpaddr, mrtaddr, vifaddr)
 			printf("v_lcl_grps_n: malloc failed\n");
 			return;
 		}
-		kread((off_t)v->v_lcl_grps, (caddr_t)grp, n * sizeof(*grp));
+		kread((u_long)v->v_lcl_grps, (caddr_t)grp, n * sizeof(*grp));
 		for (i = 0; i < n; ++i)
 			printf("%51s %-15.15s\n",
 			    "", routename((grp++)->s_addr));
@@ -113,7 +113,7 @@ mroutepr(mrpaddr, mrtaddr, vifaddr)
 	if (!banner_printed)
 		printf("\nVirtual Interface Table is empty\n");
 
-	kread((off_t)mrtaddr, (char *)&mrttable, sizeof(mrttable));
+	kread(mrtaddr, (char *)&mrttable, sizeof(mrttable));
 	banner_printed = 0;
 	for (i = 0; i < MRTHASHSIZ; ++i) {
 		for (mrt = mrttable[i]; mrt != NULL; mrt = mrt->mrt_next) {
@@ -123,7 +123,7 @@ mroutepr(mrpaddr, mrtaddr, vifaddr)
 				banner_printed = 1;
 			}
 
-			kread((off_t)mrt, (char *)&smrt, sizeof(*mrt));
+			kread((u_long)mrt, (char *)&smrt, sizeof(*mrt));
 			mrt = &smrt;
 			printf(" %3u   %-15.15s  %2u   ",
 			    i, netname(mrt->mrt_origin.s_addr,
@@ -148,7 +148,7 @@ mroutepr(mrpaddr, mrtaddr, vifaddr)
 
 void
 mrt_stats(mrpaddr, mstaddr)
-	off_t mrpaddr, mstaddr;
+	u_long mrpaddr, mstaddr;
 {
 	u_int mrtproto;
 	struct mrtstat mrtstat;
@@ -158,7 +158,7 @@ mrt_stats(mrpaddr, mstaddr)
 		return;
 	}
 
-	kread((off_t)mrpaddr, (char *)&mrtproto, sizeof(mrtproto));
+	kread(mrpaddr, (char *)&mrtproto, sizeof(mrtproto));
 	switch (mrtproto) {
 	    case 0:
 		printf("no multicast routing compiled into this system\n");
@@ -177,7 +177,7 @@ mrt_stats(mrpaddr, mstaddr)
 		return;
 	}
 
-	kread((off_t)mstaddr, (char *)&mrtstat, sizeof(mrtstat));
+	kread(mstaddr, (char *)&mrtstat, sizeof(mrtstat));
 	printf("multicast routing:\n");
 	printf(" %10u multicast route lookup%s\n",
 	  mrtstat.mrts_mrt_lookups, plural(mrtstat.mrts_mrt_lookups));
